Remove the subscription usage-limits feature
`claude setup-token` mints an inference-scoped OAuth token that the usage
endpoint rejects (`403 permission_error: OAuth token does not meet scope
requirement user:profile`). Only the interactive-login session token carries
that scope, and reading it fresh means Claude Code's Keychain item — which
re-prompts for authorization unless the app is Developer-ID signed. With no
prompt-free path to a valid token, the feature isn't worth its friction.

Remove it entirely: UsageStore, ClaudeUsageLoader, PoirotTokenStore, the
UsageLoading protocol, UsageRefreshInterval, the ClaudeUsage subscription model,
the gauge/opt-in/disclaimer cards, the dashboard usage section, the menu-bar
gauges, the Settings > Usage tab, and the `.usage` provider capability — plus
all their tests and mocks.

Kept UsageWindow / UsageSeverity / UsageGaugeRing, which the per-turn
context-window card (ConversationContextCard) still uses. Tooltips (Phase 1) and
memory folders (Phase 3) are unaffected.
