+++ Poirot/Sources/App/PoirotApp.swift
@@ -29,8 +29,6 @@ struct PoirotApp: App {
     @State
     private var appState = AppState()
     @State
-    private var usageStore = UsageStore()
-    @State
     private var memorySources = MemorySourcesStore()
     @AppStorage("showMenuBarIcon")
     private var showMenuBarIcon = true
@@ -40,7 +38,6 @@ struct PoirotApp: App {
         WindowGroup(id: "main") {
             ContentView()
                 .environment(appState)
-                .environment(usageStore)
                 .environment(memorySources)
         }
         .windowStyle(.hiddenTitleBar)
@@ -121,14 +118,12 @@ struct PoirotApp: App {
         Settings {
             SettingsView()
                 .environment(appState)
-                .environment(usageStore)
                 .environment(memorySources)
         }
 
         MenuBarExtra("Poirot", image: "MenuBarIcon", isInserted: $showMenuBarIcon) {
             MenuBarView()
                 .environment(appState)
-                .environment(usageStore)
         }
         .menuBarExtraStyle(.window)
     }
