+++ Poirot/Sources/Models/Provider/ProviderCapability.swift
@@ -9,5 +9,4 @@ enum ProviderCapability: String, CaseIterable, Sendable {
     case outputStyles
     case memory
     case hooks
-    case usage
 }
