+++ Poirot/Sources/Views/Analytics/AnalyticsDashboardView.swift
@@ -2,12 +2,8 @@ import Charts
 import SwiftUI
 
 struct AnalyticsDashboardView: View {
-    @Environment(UsageStore.self)
-    private var usageStore
     @Environment(\.provider)
     private var provider
-    @Environment(\.openSettings)
-    private var openSettings
     @State
     private var viewModel: AnalyticsViewModel
     @State
@@ -23,13 +19,8 @@ struct AnalyticsDashboardView: View {
     @State
     private var modelSelectedAngle: Int?
 
-    /// Reference time for usage reset countdowns. `nil` (the default) uses the live clock
-    /// at render time; tests inject a fixed value for deterministic snapshots.
-    private let usageNow: Date?
-
-    init(viewModel: AnalyticsViewModel = AnalyticsViewModel(), usageNow: Date? = nil) {
+    init(viewModel: AnalyticsViewModel = AnalyticsViewModel()) {
         _viewModel = State(initialValue: viewModel)
-        self.usageNow = usageNow
     }
 
     var body: some View {
@@ -54,9 +45,6 @@ struct AnalyticsDashboardView: View {
                 await viewModel.loadStats()
             }
         }
-        // Usage is never fetched on appear — only on an explicit action (Enable / Load /
-        // Refresh) or the background poll. Restarts show the persisted snapshot, so navigating
-        // here never triggers a request.
     }
 
     // MARK: - Toolbar
@@ -190,9 +178,6 @@ struct AnalyticsDashboardView: View {
         Button {
             Task {
                 await viewModel.loadStats()
-                if provider.supports(.usage) {
-                    await usageStore.refresh(force: true)
-                }
             }
         } label: {
             Image(systemName: "arrow.clockwise")
@@ -205,18 +190,14 @@ struct AnalyticsDashboardView: View {
 
     // MARK: - Dashboard Content
 
-    /// Always renders the header and the usage section (when supported), so usage limits are
-    /// reachable even when there's no stats cache. The stats area below shows charts when the
-    /// cache is present, or a compact notice when it isn't.
+    /// Always renders the header, so the dashboard is reachable even when there's no stats
+    /// cache. The stats area below shows charts when the cache is present, or a compact notice
+    /// when it isn't.
     private var dashboardContent: some View {
         VStack(spacing: 0) {
             header(viewModel.stats)
             ScrollView {
                 VStack(alignment: .leading, spacing: PoirotTheme.Spacing.xxl) {
-                    if provider.supports(.usage) {
-                        usageSection
-                    }
-
                     if let stats = viewModel.stats {
                         statsCharts(stats)
                     } else {
@@ -352,152 +333,6 @@ struct AnalyticsDashboardView: View {
             Divider().opacity(0.3)
         }
     }
-
-    // MARK: - Usage Limits
-
-    @ViewBuilder
-    private var usageSection: some View {
-        VStack(alignment: .leading, spacing: PoirotTheme.Spacing.md) {
-            HStack(spacing: PoirotTheme.Spacing.sm) {
-                Text("Usage Limits")
-                    .font(PoirotTheme.Typography.headingSmall)
-                    .foregroundStyle(PoirotTheme.Colors.textPrimary)
-                // swiftlint:disable:next line_length
-                InfoTooltipButton(text: "Your Claude subscription's rate-limit windows. Poirot uses the token you generate with `claude setup-token` — no extra login.")
-                Spacer()
-            }
-
-            usageContent
-        }
-    }
-
-    @ViewBuilder
-    private var usageContent: some View {
-        if !usageStore.isEnabled {
-            UsageOptInCard {
-                Task { await usageStore.enable() }
-            }
-        } else {
-            enabledUsageContent
-        }
-    }
-
-    @ViewBuilder
-    private var enabledUsageContent: some View {
-        switch usageStore.state {
-        case let .loaded(usage):
-            loadedUsage(usage)
-
-        case .loading:
-            UsagePlaceholderCard(
-                icon: "gauge.with.dots.needle.bottom.50percent",
-                message: "Loading usage limits…",
-                pulse: true
-            )
-
-        case let .idle(note):
-            if usageStore.hasToken {
-                // A token is stored: the tap loads/retries. `note` explains a prior failure.
-                UsageOptInCard(
-                    icon: note == nil ? "gauge.with.dots.needle.bottom.50percent" : "key.slash",
-                    title: note == nil ? "Load your usage limits" : "Couldn't read your usage limits",
-                    message: note ?? "Fetch your current 5-hour and 7-day rate-limit windows.",
-                    actionTitle: note == nil ? "Load" : "Try again",
-                    actionIcon: note == nil ? "arrow.down.circle" : "arrow.clockwise"
-                ) { Task { await usageStore.refresh(force: true) } }
-            } else {
-                // No token yet: retrying is pointless — send the user to Settings to add one.
-                UsageOptInCard(
-                    icon: "key.slash",
-                    title: "Add your usage token",
-                    // swiftlint:disable:next line_length
-                    message: note ?? "Run `claude setup-token` in your terminal, then paste the token in Settings › Usage.",
-                    actionTitle: "Open Settings",
-                    actionIcon: "gearshape"
-                ) { openSettings() }
-            }
-        }
-    }
-
-    @ViewBuilder
-    private func loadedUsage(_ usage: ClaudeUsage) -> some View {
-        let now = usageNow ?? Date()
-        VStack(spacing: PoirotTheme.Spacing.md) {
-            HStack(spacing: PoirotTheme.Spacing.md) {
-                UsageGaugeCard(title: "5-Hour Window", icon: "clock.arrow.circlepath", window: usage.fiveHour, now: now)
-                UsageGaugeCard(title: "7-Day Window", icon: "calendar", window: usage.sevenDay, now: now)
-            }
-            .frame(height: usageCardHeight)
-
-            if usage.sevenDayOpus != nil || usage.sevenDaySonnet != nil {
-                HStack(spacing: PoirotTheme.Spacing.md) {
-                    if let opus = usage.sevenDayOpus {
-                        UsageGaugeCard(title: "Opus · Weekly", icon: "sparkle", window: opus, now: now)
-                    }
-                    if let sonnet = usage.sevenDaySonnet {
-                        UsageGaugeCard(title: "Sonnet · Weekly", icon: "wand.and.stars", window: sonnet, now: now)
-                    }
-                }
-                .frame(height: usageCardHeight)
-            }
-
-            if let spend = usage.spend, spend.enabled {
-                UsagePlaceholderCard(
-                    icon: "dollarsign.circle",
-                    message: "Spend: \(String(format: "$%.2f", spend.usedDollars)) · \(Int(spend.utilization.rounded()))% of limit",
-                    tint: PoirotTheme.Colors.green
-                )
-            }
-
-            if let extra = usage.extraUsage, extra.enabled {
-                UsagePlaceholderCard(
-                    icon: "plus.circle",
-                    message: "Extra usage: \(Int((extra.utilization ?? 0).rounded()))% used",
-                    tint: PoirotTheme.Colors.orange
-                )
-            }
-
-            usageStatusFooter(now: now)
-        }
-    }
-
-    /// A one-line footer under the gauges: when Anthropic is throttling us, a countdown and a
-    /// disclaimer; otherwise how fresh the data is and when the next auto-refresh lands.
-    @ViewBuilder
-    private func usageStatusFooter(now: Date) -> some View {
-        HStack(spacing: PoirotTheme.Spacing.xs) {
-            if let until = usageStore.rateLimitedUntil, until > now {
-                Image(systemName: "clock.badge.exclamationmark")
-                    .foregroundStyle(PoirotTheme.Colors.orange)
-                // swiftlint:disable:next line_length
-                Text("Anthropic limits how often usage can be checked — retrying in \(UsageCountdown.format(until.timeIntervalSince(now)))")
-                    .foregroundStyle(PoirotTheme.Colors.textTertiary)
-            } else {
-                Image(systemName: "checkmark.seal")
-                    .foregroundStyle(PoirotTheme.Colors.textTertiary)
-                Text(usageFreshnessText(now: now))
-                    .foregroundStyle(PoirotTheme.Colors.textTertiary)
-            }
-            Spacer(minLength: 0)
-        }
-        .font(PoirotTheme.Typography.micro)
-    }
-
-    private func usageFreshnessText(now: Date) -> String {
-        var parts: [String] = []
-        if let last = usageStore.lastUpdated {
-            parts.append("Updated \(UsageCountdown.format(now.timeIntervalSince(last))) ago")
-        }
-        if let next = usageStore.nextRefreshAt {
-            let remaining = next.timeIntervalSince(now)
-            parts.append(remaining > 0 ? "next in \(UsageCountdown.format(remaining))" : "next update due")
-        } else if usageStore.refreshInterval == .manual {
-            parts.append("auto-refresh off")
-        }
-        return parts.joined(separator: " · ")
-    }
-
-    private let usageCardHeight: CGFloat = 96
 
     // MARK: - Summary Cards
