+++ Poirot/Sources/Views/MenuBar/MenuBarView.swift
@@ -3,8 +3,6 @@ import SwiftUI
 struct MenuBarView: View {
     @Environment(AppState.self)
     private var appState
-    @Environment(UsageStore.self)
-    private var usageStore
     @Environment(\.provider)
     private var provider
     @Environment(\.openWindow)
@@ -15,10 +13,6 @@ struct MenuBarView: View {
     var body: some View {
         VStack(spacing: 0) {
             headerSection
-            if showUsageSection {
-                Divider().opacity(0.3)
-                usageSection
-            }
             Divider().opacity(0.3)
             searchField
             Divider().opacity(0.3)
@@ -32,77 +26,6 @@ struct MenuBarView: View {
             menuBarState.loadRecentSessions(from: appState.projects)
             menuBarState.loadStats()
         }
-        // The menu bar never fetches usage itself (no Keychain read on open) — it passively
-        // mirrors whatever the dashboard has already loaded this session.
-    }
-
-    // MARK: - Usage
-
-    private var showUsageSection: Bool {
-        guard provider.supports(.usage), usageStore.isEnabled else { return false }
-        if case .loaded = usageStore.state { return true }
-        return false
-    }
-
-    @ViewBuilder
-    private var usageSection: some View {
-        if case let .loaded(usage) = usageStore.state {
-            VStack(alignment: .leading, spacing: PoirotTheme.Spacing.xs) {
-                usageHeader
-                MenuBarUsageRow(label: "5h", window: usage.fiveHour)
-                MenuBarUsageRow(label: "7d", window: usage.sevenDay)
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .padding(.horizontal, PoirotTheme.Spacing.lg)
-            .padding(.vertical, PoirotTheme.Spacing.sm)
-        }
-    }
-
-    private var usageHeader: some View {
-        HStack(spacing: PoirotTheme.Spacing.xs) {
-            Text("Usage")
-                .font(PoirotTheme.Typography.microSemibold)
-                .foregroundStyle(PoirotTheme.Colors.textTertiary)
-
-            Spacer()
-
-            usageCadenceMenu
-
-            Button {
-                Task { await usageStore.refresh(force: true) }
-            } label: {
-                Image(systemName: "arrow.clockwise")
-                    .font(PoirotTheme.Typography.tiny)
-                    .foregroundStyle(PoirotTheme.Colors.textSecondary)
-            }
-            .buttonStyle(.plain)
-            .disabled(usageStore.state == .loading)
-            .help("Refresh usage now")
-        }
-    }
-
-    private var usageCadenceMenu: some View {
-        Menu {
-            Picker("Auto-refresh", selection: Binding(
-                get: { usageStore.refreshInterval },
-                set: { usageStore.setRefreshInterval($0) }
-            )) {
-                ForEach(UsageRefreshInterval.allCases) { option in
-                    Text(option.label).tag(option)
-                }
-            }
-        } label: {
-            HStack(spacing: 2) {
-                Image(systemName: usageStore.refreshInterval == .manual ? "hand.tap" : "timer")
-                Text(usageStore.refreshInterval.shortLabel)
-            }
-            .font(PoirotTheme.Typography.tiny)
-            .foregroundStyle(PoirotTheme.Colors.textTertiary)
-        }
-        .menuStyle(.borderlessButton)
-        .menuIndicator(.hidden)
-        .fixedSize()
-        .help("Auto-refresh interval")
     }
 
     // MARK: - Header
@@ -269,51 +192,6 @@ struct MenuBarView: View {
         appState.selectedProject = projectId
         appState.selectedSession = session
         appState.selectedNav = .sessions
-    }
-}
-
-// MARK: - Usage Row
-
-private struct MenuBarUsageRow: View {
-    let label: String
-    let window: UsageWindow
-    var now = Date()
-
-    var body: some View {
-        HStack(spacing: PoirotTheme.Spacing.sm) {
-            Text(label)
-                .font(PoirotTheme.Typography.microSemibold)
-                .foregroundStyle(PoirotTheme.Colors.textSecondary)
-                .frame(width: 18, alignment: .leading)
-
-            GeometryReader { geo in
-                ZStack(alignment: .leading) {
-                    Capsule()
-                        .fill(PoirotTheme.Colors.bgElevated)
-                    Capsule()
-                        .fill(window.severity.tint)
-                        .frame(width: max(3, geo.size.width * window.fraction))
-                }
-            }
-            .frame(height: 5)
-
-            Text("\(window.percent)%")
-                .font(PoirotTheme.Typography.microMedium)
-                .foregroundStyle(PoirotTheme.Colors.textPrimary)
-                .frame(width: 32, alignment: .trailing)
-
-            Text(resetText)
-                .font(PoirotTheme.Typography.tiny)
-                .foregroundStyle(PoirotTheme.Colors.textTertiary)
-                .frame(width: 58, alignment: .trailing)
-        }
-    }
-
-    private var resetText: String {
-        guard let interval = window.timeUntilReset(now: now), interval >= 60 else {
-            return "resets now"
-        }
-        return UsageCountdown.format(interval)
     }
 }
