--- PoirotTests/App/UsageStoreTests.swift
@@ -1,457 +1,0 @@
-@testable import Poirot
-import Foundation
-import Testing
-
-@MainActor
-@Suite("UsageStore")
-struct UsageStoreTests {
-    private static let sampleUsage = ClaudeUsage(
-        fiveHour: UsageWindow(utilization: 20, resetsAt: nil),
-        sevenDay: UsageWindow(utilization: 30, resetsAt: nil),
-        sevenDayOpus: nil,
-        sevenDaySonnet: nil,
-        spend: nil,
-        extraUsage: nil
-    )
-
-    /// A throwaway defaults suite so opt-in/snapshot persistence never touches `.standard` or
-    /// leaks between tests.
-    private static func freshDefaults(enabled: Bool) -> UserDefaults {
-        let defaults = UserDefaults(suiteName: "UsageStoreTests-\(UUID().uuidString)")!
-        defaults.set(enabled, forKey: UsageStore.enabledKey)
-        return defaults
-    }
-
-    private static func isIdle(_ state: UsageStore.State) -> Bool {
-        if case .idle = state { return true }
-        return false
-    }
-
-    /// Polls `condition` until it holds or the ceiling elapses, yielding between checks. Keeps
-    /// timer-based tests robust under heavy parallel load instead of depending on a fixed sleep.
-    private static func waitUntil(
-        timeout: Duration = .seconds(5),
-        _ condition: () -> Bool
-    ) async throws {
-        let deadline = ContinuousClock.now + timeout
-        while !condition(), ContinuousClock.now < deadline {
-            try await Task.sleep(for: .milliseconds(10))
-        }
-    }
-
-    // MARK: - Opt-in gating
-
-    @Test
-    func refresh_whenDisabled_doesNotFetch() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let store = UsageStore(loader: mock, throttle: 0, defaults: Self.freshDefaults(enabled: false))
-
-        await store.refresh(force: true)
-
-        #expect(!store.isEnabled)
-        #expect(mock.loadUsageCallsCount == 0)
-        #expect(Self.isIdle(store.state))
-    }
-
-    @Test
-    func enable_fetchesAndPersists() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let defaults = Self.freshDefaults(enabled: false)
-        let store = UsageStore(loader: mock, throttle: 0, defaults: defaults)
-
-        await store.enable()
-
-        #expect(store.isEnabled)
-        #expect(mock.loadUsageCallsCount == 1)
-        #expect(store.state == .loaded(Self.sampleUsage))
-        #expect(defaults.bool(forKey: UsageStore.enabledKey))
-    }
-
-    // MARK: - Persistence (rule 2: no Keychain read on restart)
-
-    @Test
-    func restoresSnapshotOnInit_withoutFetching() async {
-        let defaults = Self.freshDefaults(enabled: true)
-        let writer = UsageLoadingMock()
-        writer.loadUsageReturnValue = .success(Self.sampleUsage)
-        await UsageStore(loader: writer, throttle: 0, defaults: defaults).refresh()
-
-        // A brand-new store (simulating relaunch) restores the snapshot and never calls the loader.
-        let reader = UsageLoadingMock()
-        let restored = UsageStore(loader: reader, throttle: 0, defaults: defaults)
-
-        #expect(restored.usage == Self.sampleUsage)
-        #expect(reader.loadUsageCallsCount == 0)
-    }
-
-    @Test
-    func disabledStore_doesNotRestoreSnapshot() async {
-        let defaults = Self.freshDefaults(enabled: true)
-        let writer = UsageLoadingMock()
-        writer.loadUsageReturnValue = .success(Self.sampleUsage)
-        await UsageStore(loader: writer, throttle: 0, defaults: defaults).refresh()
-
-        // Opt back out → snapshot is forgotten and not restored.
-        defaults.set(false, forKey: UsageStore.enabledKey)
-        let restored = UsageStore(loader: UsageLoadingMock(), throttle: 0, defaults: defaults)
-
-        #expect(restored.usage == nil)
-    }
-
-    // MARK: - Fetch outcomes
-
-    @Test
-    func refresh_success_setsLoaded() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let store = UsageStore(loader: mock, throttle: 0, defaults: Self.freshDefaults(enabled: true))
-
-        await store.refresh()
-
-        #expect(store.state == .loaded(Self.sampleUsage))
-        #expect(store.usage == Self.sampleUsage)
-    }
-
-    @Test
-    func refresh_unauthenticated_clearsSnapshotAndGoesIdle() async {
-        let defaults = Self.freshDefaults(enabled: true)
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let store = UsageStore(loader: mock, throttle: 0, defaults: defaults)
-        await store.refresh()
-
-        mock.loadUsageReturnValue = .unauthenticated
-        await store.refresh(force: true)
-
-        guard case let .idle(note) = store.state else {
-            Issue.record("expected idle after unauthenticated")
-            return
-        }
-        #expect(note != nil)
-        // Snapshot was cleared, so a relaunch shows the Load state rather than stale gauges.
-        let restored = UsageStore(loader: UsageLoadingMock(), throttle: 0, defaults: defaults)
-        #expect(restored.usage == nil)
-    }
-
-    // MARK: - Token
-
-    @Test
-    func saveToken_setsHasToken_andLoadsWhenEnabled() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let tokens = OAuthTokenStoringMock()
-        let store = UsageStore(
-            loader: mock,
-            tokenStore: tokens,
-            throttle: 0,
-            autoRefreshEnabled: false,
-            defaults: Self.freshDefaults(enabled: true)
-        )
-        #expect(!store.hasToken)
-
-        await store.saveToken("  sk-ant-oat-xyz  ")
-
-        #expect(store.hasToken)
-        #expect(tokens.token == "sk-ant-oat-xyz") // trimmed
-        #expect(mock.loadUsageCalled)
-        #expect(store.usage != nil)
-    }
-
-    @Test
-    func saveToken_whenBlank_isIgnored() async {
-        let tokens = OAuthTokenStoringMock()
-        let store = UsageStore(
-            loader: UsageLoadingMock(),
-            tokenStore: tokens,
-            throttle: 0,
-            autoRefreshEnabled: false,
-            defaults: Self.freshDefaults(enabled: true)
-        )
-
-        await store.saveToken("   ")
-
-        #expect(!store.hasToken)
-        #expect(tokens.token == nil)
-    }
-
-    @Test
-    func clearToken_forgetsTokenAndGoesIdle() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let tokens = OAuthTokenStoringMock(token: "sk-existing")
-        let store = UsageStore(
-            loader: mock,
-            tokenStore: tokens,
-            throttle: 0,
-            autoRefreshEnabled: false,
-            defaults: Self.freshDefaults(enabled: true)
-        )
-        await store.refresh(force: true)
-        #expect(store.hasToken)
-
-        store.clearToken()
-
-        #expect(!store.hasToken)
-        #expect(tokens.token == nil)
-        #expect(Self.isIdle(store.state))
-    }
-
-    @Test
-    func unauthenticated_withoutToken_promptsToAddOne() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .unauthenticated
-        let store = UsageStore(
-            loader: mock,
-            tokenStore: OAuthTokenStoringMock(token: nil),
-            throttle: 0,
-            autoRefreshEnabled: false,
-            defaults: Self.freshDefaults(enabled: true)
-        )
-
-        await store.refresh(force: true)
-
-        guard case let .idle(note) = store.state else {
-            Issue.record("expected idle after unauthenticated")
-            return
-        }
-        #expect(note?.contains("Add your usage token") == true)
-    }
-
-    @Test
-    func refresh_failureFromIdle_goesIdleWithNote() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .failure
-        let store = UsageStore(loader: mock, throttle: 0, defaults: Self.freshDefaults(enabled: true))
-
-        await store.refresh()
-
-        guard case let .idle(note) = store.state else {
-            Issue.record("expected idle after failure")
-            return
-        }
-        #expect(note != nil)
-    }
-
-    @Test
-    func refresh_rateLimited_suppressesEvenForcedDuringWindow() async {
-        let mock = UsageLoadingMock()
-        // A comfortably long server backoff so the window is unambiguously active.
-        mock.loadUsageReturnValue = .rateLimited(retryAfter: 60)
-        let store = UsageStore(
-            loader: mock, throttle: 0, autoRefreshEnabled: false, defaults: Self.freshDefaults(enabled: true)
-        )
-
-        await store.refresh(force: true) // hits 429 → arms the 60s backoff
-        #expect(mock.loadUsageCallsCount == 1)
-        #expect(store.rateLimitedUntil != nil)
-
-        await store.refresh() // timer tick suppressed during backoff
-        #expect(mock.loadUsageCallsCount == 1)
-
-        // An explicit tap is ALSO suppressed inside the window — a request now just earns
-        // another 429 and can extend the lockout.
-        await store.refresh(force: true)
-        #expect(mock.loadUsageCallsCount == 1)
-    }
-
-    @Test
-    func refresh_rateLimitedFromIdle_showsNote() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .rateLimited(retryAfter: nil)
-        let store = UsageStore(
-            loader: mock, throttle: 0, autoRefreshEnabled: false, defaults: Self.freshDefaults(enabled: true)
-        )
-
-        await store.refresh(force: true)
-
-        guard case let .idle(note) = store.state else {
-            Issue.record("expected idle after rate limit")
-            return
-        }
-        #expect(note?.contains("limiting") == true)
-    }
-
-    @Test
-    func refresh_rateLimitedAfterLoaded_keepsStaleData() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let store = UsageStore(
-            loader: mock, throttle: 0, autoRefreshEnabled: false, defaults: Self.freshDefaults(enabled: true)
-        )
-        await store.refresh()
-
-        mock.loadUsageReturnValue = .rateLimited(retryAfter: nil)
-        await store.refresh(force: true)
-
-        // Rate limiting keeps the cached gauges instead of dropping to an error state.
-        #expect(store.usage == Self.sampleUsage)
-    }
-
-    @Test
-    func refresh_networkFailureAfterLoaded_keepsStaleData() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let store = UsageStore(loader: mock, throttle: 0, defaults: Self.freshDefaults(enabled: true))
-        await store.refresh()
-
-        mock.loadUsageReturnValue = .failure
-        await store.refresh(force: true)
-
-        // Transient network failure keeps the cached gauges rather than dropping to Load.
-        #expect(store.usage == Self.sampleUsage)
-    }
-
-    @Test
-    func disable_clearsStateAndStops() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let store = UsageStore(loader: mock, throttle: 0, defaults: Self.freshDefaults(enabled: true))
-        await store.refresh()
-
-        store.disable()
-
-        #expect(!store.isEnabled)
-        #expect(Self.isIdle(store.state))
-    }
-
-    // MARK: - Throttle
-
-    @Test
-    func refresh_withinThrottleWindow_skipsSecondFetch() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let store = UsageStore(loader: mock, throttle: 1000, defaults: Self.freshDefaults(enabled: true))
-
-        await store.refresh()
-        await store.refresh()
-
-        #expect(mock.loadUsageCallsCount == 1)
-    }
-
-    @Test
-    func refresh_force_bypassesThrottle() async {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let store = UsageStore(loader: mock, throttle: 1000, defaults: Self.freshDefaults(enabled: true))
-
-        await store.refresh()
-        await store.refresh(force: true)
-
-        #expect(mock.loadUsageCallsCount == 2)
-    }
-
-    @Test
-    func preview_pinsStateWithoutFetching() async {
-        let store = UsageStore.preview(.loaded(Self.sampleUsage))
-
-        await store.refresh()
-
-        #expect(store.isEnabled)
-        #expect(store.usage == Self.sampleUsage)
-    }
-
-    // MARK: - Background auto-refresh
-
-    @Test
-    func autoRefresh_pollsWhileLoaded() async throws {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let store = UsageStore(
-            loader: mock,
-            throttle: 0,
-            autoRefreshOverride: 0.02,
-            defaults: Self.freshDefaults(enabled: false)
-        )
-
-        await store.enable() // explicit first read (count == 1), then the timer takes over
-
-        // Wait until the background timer fetches again, with a generous ceiling so the test
-        // stays robust under load rather than depending on an exact number of ticks in 0.2s.
-        try await Self.waitUntil { mock.loadUsageCallsCount > 1 }
-
-        #expect(mock.loadUsageCallsCount > 1)
-        store.stopAutoRefresh()
-    }
-
-    @Test
-    func autoRefresh_doesNotPollFromIdle() async throws {
-        // Opted in but never loaded (no snapshot) → the timer runs but must not read from idle.
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let store = UsageStore(
-            loader: mock,
-            throttle: 0,
-            autoRefreshOverride: 0.02,
-            defaults: Self.freshDefaults(enabled: true)
-        )
-
-        #expect(Self.isIdle(store.state))
-        try await Task.sleep(for: .seconds(0.2))
-
-        #expect(mock.loadUsageCallsCount == 0)
-        store.stopAutoRefresh()
-    }
-
-    @Test
-    func disable_stopsAutoRefresh() async throws {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let store = UsageStore(
-            loader: mock,
-            throttle: 0,
-            autoRefreshOverride: 0.02,
-            defaults: Self.freshDefaults(enabled: false)
-        )
-        await store.enable()
-
-        store.disable()
-        try await Task.sleep(for: .seconds(0.1)) // let any in-flight tick settle
-        let countAtDisable = mock.loadUsageCallsCount
-        try await Task.sleep(for: .seconds(0.2))
-
-        // No further fetches after opting out.
-        #expect(mock.loadUsageCallsCount == countAtDisable)
-    }
-
-    // MARK: - Refresh cadence
-
-    @Test
-    func refreshInterval_defaultsAndPersists() {
-        let defaults = Self.freshDefaults(enabled: false)
-        let store = UsageStore(loader: UsageLoadingMock(), autoRefreshEnabled: false, defaults: defaults)
-
-        #expect(store.refreshInterval == .default)
-
-        store.setRefreshInterval(.tenMinutes)
-        #expect(store.refreshInterval == .tenMinutes)
-        #expect(defaults.integer(forKey: UsageStore.intervalKey) == 10)
-
-        // A relaunch restores the chosen cadence.
-        let restored = UsageStore(loader: UsageLoadingMock(), autoRefreshEnabled: false, defaults: defaults)
-        #expect(restored.refreshInterval == .tenMinutes)
-    }
-
-    @Test
-    func setRefreshInterval_manual_stopsPolling() async throws {
-        let mock = UsageLoadingMock()
-        mock.loadUsageReturnValue = .success(Self.sampleUsage)
-        let store = UsageStore(
-            loader: mock,
-            throttle: 0,
-            autoRefreshOverride: 0.02,
-            defaults: Self.freshDefaults(enabled: false)
-        )
-        await store.enable() // polling on the default (automatic) cadence
-        try await Self.waitUntil { mock.loadUsageCallsCount > 1 }
-
-        store.setRefreshInterval(.manual)
-        try await Task.sleep(for: .seconds(0.1)) // let any in-flight tick settle
-        let settled = mock.loadUsageCallsCount
-        try await Task.sleep(for: .seconds(0.2))
-
-        // Manual mode halts the background poll; the count no longer grows on its own.
-        #expect(mock.loadUsageCallsCount == settled)
-    }
-}
