+++ PoirotTests/Screenshots/Helpers/ScreenshotTestHelpers.swift
@@ -159,14 +159,12 @@ func withEnvironment<V: View>(
     state: AppState? = nil,
     provider: any ProviderDescribing = ClaudeCodeProvider(),
     historyLoader: (any HistoryLoading)? = nil,
-    usageStore: UsageStore = .preview(.idle(note: nil)),
     memorySources: MemorySourcesStore = .previewEmpty()
 ) -> some View {
     let appState = state ?? makeAppState()
     let result = view
         .environment(appState)
         .environment(\.provider, provider)
-        .environment(usageStore)
         .environment(memorySources)
     if let historyLoader {
         return AnyView(result.environment(\.historyLoader, historyLoader))
@@ -183,7 +181,6 @@ func compositeAppView<Detail: View>(
     state: AppState? = nil,
     provider: any ProviderDescribing = ClaudeCodeProvider(),
     historyLoader: (any HistoryLoading)? = nil,
-    usageStore: UsageStore = .preview(.idle(note: nil)),
     memorySources: MemorySourcesStore = .previewEmpty(),
     @ViewBuilder detail: () -> Detail
 ) -> some View {
@@ -198,7 +195,6 @@ func compositeAppView<Detail: View>(
     }
     .environment(appState)
     .environment(\.provider, provider)
-    .environment(usageStore)
     .environment(memorySources)
 
     if let historyLoader {
